+++ CopilotKit/packages/runtime/src/lib/runtime/remote-actions.ts
@@ -1,4 +1,4 @@
-import { Action } from "@copilotkit/shared";
+import { Action, CopilotKitErrorCode } from "@copilotkit/shared";
 import { GraphQLContext } from "../integrations/shared";
 import { Logger } from "pino";
 import { Message } from "../../graphql/types/converted";
@@ -17,6 +17,8 @@ import {
   CopilotKitError,
 } from "@copilotkit/shared";
 import { MetaEventInput } from "../../graphql/inputs/meta-event.input";
+import { AbstractAgent } from "@agentwire/client";
+import { constructAgentWireRemoteAction } from "./agentwire-action";
 
 export type EndpointDefinition = CopilotKitEndpoint | LangGraphPlatformEndpoint;
 
@@ -127,12 +129,14 @@ export async function setupRemoteActions({
   messages,
   agentStates,
   frontendUrl,
+  agents,
 }: {
   remoteEndpointDefinitions: EndpointDefinition[];
   graphqlContext: GraphQLContext;
   messages: Message[];
   agentStates?: AgentStateInput[];
   frontendUrl?: string;
+  agents: Record<string, AbstractAgent>;
 }): Promise<Action[]> {
   const logger = graphqlContext.logger.child({ component: "remote-actions.fetchRemoteActions" });
   logger.debug({ remoteEndpointDefinitions }, "Fetching from remote endpoints");
@@ -180,6 +184,26 @@ export async function setupRemoteActions({
       });
     }),
   );
+
+  for (const [key, agent] of Object.entries(agents)) {
+    if (agent.agentId !== undefined && agent.agentId !== key) {
+      throw new CopilotKitError({
+        message: `Agent ${key} has agentId ${agent.agentId} which does not match the key ${key}`,
+        code: CopilotKitErrorCode.UNKNOWN,
+      });
+    } else if (agent.agentId === undefined) {
+      agent.agentId = key;
+    }
+
+    result.push(
+      constructAgentWireRemoteAction({
+        logger,
+        messages,
+        agentStates,
+        agent: agent,
+      }),
+    );
+  }
 
   return result.flat();
 }