+++ CopilotKit/packages/react-ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # ui
 
+## 1.8.5-next.5
+
+### Patch Changes
+
+- Updated dependencies [c0d3261]
+  - @copilotkit/react-core@1.8.5-next.5
+  - @copilotkit/runtime-client-gql@1.8.5-next.5
+  - @copilotkit/shared@1.8.5-next.5
+
+## 1.8.5-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.5-next.4
+- @copilotkit/react-core@1.8.5-next.4
+- @copilotkit/shared@1.8.5-next.4
+
 ## 1.8.5-next.3
 
 ### Patch Changes