+++ CopilotKit/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkit/sdk-js
 
+## 1.8.5-next.5
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.5-next.5
+
+## 1.8.5-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.5-next.4
+
 ## 1.8.5-next.3
 
 ### Patch Changes