+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @copilotkit/shared
 
+## 1.8.5-next.5
+
+## 1.8.5-next.4
+
 ## 1.8.5-next.3
 
 ## 1.8.5-next.2